requirements.txt
asttokens==3.0.0
comm==0.2.2
config_loader==1.0.0
contourpy==1.3.0
cupy-cuda12x==13.4.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
exceptiongroup==1.2.2
executing==2.2.0
fastrlock==0.8.3
ffmpeg-python==0.2.0
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.2
future==1.0.0
imageio==2.36.0
imageio-ffmpeg==0.5.1
importlib_metadata==8.6.1
importlib_resources==6.5.2
intel-cmplr-lib-ur==2025.1.0
intel-openmp==2025.1.0
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.6
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.14
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mkl==2025.1.0
mkl-service==2.4.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnx==1.17.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
protobuf==6.30.2
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
PyImpetus==4.1.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-wavelets==1.3.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.3.0
ruptures==1.1.9
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
tbb==2022.1.0
tcmlib==1.3.0
threadpoolctl==3.5.0
torch==2.6.0
torchaudio==2.6.0
torchinfo==1.8.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.2
umf==0.10.0
wcwidth==0.2.13
widgetsnbextension==4.0.14
zipp==3.21.0
